Add check driver for detecting C and Ada type mismatch.
This test driver can be used to compare the size of the C
struct and the corresponding Ada record. It can also be used
to check record field offset (this can be useful with complexe
record layout).

This driver has been used to help migrating Win32Ada to Windows
64bits. Some checks are still failing.
--
Part of J323-006.