Rework script to build Win32Ada.
The template script for the installer has been removed the .in suffix
as it is not a template anymore. The GNATprep call has been removed
in Dev 2013 (985adb0) and so there is nothing edited on this file.

Also a simplified user script is added for building from sources.

TN: eng/toolchain/win32ada#6